Hw5,RockPaperScissors,DatingProgram,Turtle Art/etc
This Code has the following programs where it will ask the user to perform one of the 4 options of a Rock, Paper, Scissors Game vs computer, a Dating Program or a Turtle Artwork premade, Lastly if the user wants to exit the program. The second program is another menu with 4 options of to calculate a cube number, finding a number close 23, or to roll a random dice and n command being to quit the program.