Hw7, Using Files/.csv and Plotting bar charts
This code allows the user to edit a .csv file "monthlysales.csv" and can view the monthly,yearly sales while being able to edit the files of the sales and lastly can exit the program. The Second Program allows the user to print out a bar chart of the United States Population preset data values, using import matplotlib.pyplot as a plt, and numpy as np. Which are both essential imports for print out bar graphs.