Csv file for Hw7 Code to function properly
This is the monthly sales/yearly sales.csv file which is needed for the Hw7 Code Program to function properly. This is going to allow the user to edit/make changes in the file. If this file is not in the same location as the Hw7 Code file it will not function.